A few fixes and extensions
Here is some code extensions and bug fixes that you might be interested in. I
have also included some rebased diffs from two forks of your code on github
because they seemed interesting. One does a basic dump of returned data from
the camera and the other allows access to multiple profiles.

My code adds some extra fields that I was interested in as well as a couple of
extra calls (including get/set NTP).

I've also done some reworking of your time logic to better reflect UTC time vs.
localtime and produced a utility to do time sync to cameras.

The code is based on a patch from Brian D Scott, with some indentation cleanup.
I did not complete the indentation cleanup, as it is unclear to me what the
intended indentation for this project is.

Fixes #3.